+++ packages/renderer/src/chrome/app-shell.ts
@@ -477,7 +477,7 @@ export class AppShell extends LitElement {
             target="_blank"
             rel="noopener"
             title="View the changelog on GitHub"
-            ><span class="version">v0.0.274</span></a
+            ><span class="version">v0.0.275</span></a
           ></strong
         >
         ${this.headerButtons